Generation of deployment POMs with templates.
