+++ bw-maven-plugin/src/main/java/fr/fastconnect/factory/tibco/bw/maven/packaging/pom/GenerateDeploymentPOM.java
@@ -1,0 +1,165 @@
+/**
+ * (C) Copyright 2011-2015 FastConnect SAS
+ * (http://www.fastconnect.fr/) and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package fr.fastconnect.factory.tibco.bw.maven.packaging.pom;
+
+import java.io.File;
+import java.io.InputStream;
+
+import org.apache.maven.model.Model;
+import org.apache.maven.model.Parent;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.project.MavenProject;
+
+/**
+ * <p>
+ * This goal creates a deployment POM with
+ * "<a href="./bwmaven-lifecycles.html#Deployment">bw-ear-deploy</a>" packaging.
+ * <br/>
+ * This POM will be used at deployment time. It depends on the parent deployment
+ * POM generated by <a href="">generate-pom</a> goal.
+ * </p>
+ *
+ * @author Mathieu Debove
+ *
+ */
+@Mojo( name="generate-deployment-pom",
+defaultPhase=LifecyclePhase.PREPARE_PACKAGE )
+public class GenerateDeploymentPOM extends AbstractPOMGenerator {
+
+	@Parameter (property="deploy.pom.deployment", required=true, defaultValue="pom.xml")
+	protected String pomDeployFilename;
+
+	/**
+	 * The deployment POM template is a partial Maven POM which will be used to
+	 * generate the deployment POM.
+	 */
+	@Parameter (property="deploy.pom.deployment.template", required=false, defaultValue="${project.basedir}/src/main/maven/deployment-pom.xml")
+	protected File pomDeployTemplate;
+
+	/**
+	 * Whether to merge the deployment POM template with the built-in
+	 * deployment POM.
+	 */
+	@Parameter (property="deploy.pom.deployment.template.merge", required=false, defaultValue="true")
+	protected Boolean pomDeployTemplateMerge;
+
+	/**
+	 * To skip the generation of the POM.
+	 * This might be useful to ignore intermediate parent POM between "bw-ear"
+	 * POMs and the real top level parent POM.
+	 */
+	@Parameter ( property = "generate.pom.deployment.skip", required = false, defaultValue = "false")
+	protected boolean skipGeneratePom;
+
+	/* parent definition */
+	/**
+	 * The groupId of the parent.
+	 */
+	@Parameter ( property = "generate.pom.parent.groupId", required = true)
+	protected String parentGroupId;
+
+	/**
+	 * The artifactId of the parent.
+	 */
+	@Parameter ( property = "generate.pom.parent.artifactId", required = true)
+	protected String parentArtifactId;
+
+	/**
+	 * The version of the parent.
+	 */
+	@Parameter ( property = "generate.pom.parent.version", required = true)
+	protected String parentVersion;
+
+	/**
+	 * The suffix of artifactId of the parent.
+	 */
+	@Parameter ( property = "generate.pom.parent.suffix", required = true, defaultValue = "deploy")
+	protected String parentArtifactSuffix;
+	/* */
+
+	@Override
+	protected File getOutputFile() {
+		return new File(packageDirectory + File.separator + pomDeployFilename);
+	}
+
+	@Override
+	protected File getTemplateFile() {
+		return pomDeployTemplate;
+	}
+
+	@Override
+	protected Boolean getTemplateMerge()  {
+		return pomDeployTemplateMerge;
+	}
+
+	@Override
+	protected Boolean getSkipGeneratePOM() {
+		return skipGeneratePom;
+	}
+
+	@Override
+	protected InputStream getBuiltinTemplateFile() {
+		return this.getClass().getClassLoader().getResourceAsStream("deploy/ear/pom.xml");
+	}
+
+	@Override
+	protected Model updateModel(Model model, MavenProject project) throws MojoExecutionException {
+		model = addParent(model, parentGroupId, parentArtifactId + getArtifactSuffix(), parentVersion);
+
+		return model;
+	}
+
+	@Override
+	protected void attachFile(File pom) {
+		attachFile(pom, POM_TYPE, "deployment");
+	}
+
+	@Override
+	protected void postGeneration(File outputFile, Model model,	MavenProject project) throws MojoExecutionException {
+		// nothing to do
+	}
+
+	private Model addParent(Model model, String parentGroupId, String parentArtifactId, String parentVersion) {
+		if (model == null) return null;
+
+		Parent p = new Parent();
+		p.setGroupId(parentGroupId);
+		p.setArtifactId(parentArtifactId);
+		p.setVersion(parentVersion);
+
+		if (!(model.getGroupId()+":"+model.getArtifactId()).equals(parentGroupId+":"+parentArtifactId)) {
+			model.setParent(p);
+		}
+
+		return model;
+	}
+
+	private String getArtifactSuffix() {
+		if (parentArtifactSuffix != null && !parentArtifactSuffix.isEmpty()) {
+			if (parentArtifactSuffix.startsWith("-")) {
+				return parentArtifactSuffix;
+			} else {
+				return "-" + parentArtifactSuffix;
+			}
+		} else {
+			return "";
+		}
+	}
+}
